# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk proyek Anda.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/mahasiswa": {
            "get": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Dapatkan daftar semua mahasiswa",
                "description": "Mengembalikan daftar semua mahasiswa",
                "operationId": "getMahasiswa",
                "responses": {
                    "200": {
                        "description": "Daftar mahasiswa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mahasiswa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambahkan mahasiswa baru",
                "description": "Menambahkan mahasiswa baru ke dalam database",
                "operationId": "addMahasiswa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Mahasiswa"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mahasiswa berhasil ditambahkan"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Mahasiswa": {
                "title": "Mahasiswa",
                "description": "Model Mahasiswa",
                "properties": {
                    "id": {
                        "description": "ID mahasiswa",
                        "type": "integer"
                    },
                    "nama": {
                        "description": "Nama mahasiswa",
                        "type": "string"
                    },
                    "nim": {
                        "description": "Nomor Induk Mahasiswa",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email mahasiswa",
                        "type": "string"
                    },
                    "jurusan": {
                        "description": "Jurusan mahasiswa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Mahasiswa",
            "description": "Mahasiswa"
        }
    ]
}